refactor(racingcar) : model, service, view 함수 매개변수 수정 및 반환형 수정
- RacingCar의 getDistance 반환형 int로 수정
- WinnerEvaluator의 findMaxDistance 반환형 int로 수정
- RaceControllerFactory에서 ResultViewer 객체를 생성 후 주입하도록 수정
- ResultViewer의 output함수 println 대신 print를 사용하도록 수정

+++ src/main/java/racingcar/model/RacingCar.java
@@ -2,6 +2,6 @@ package racingcar.model;
 
 public interface RacingCar {
     String getName();
-    Long getDistance();
+    int getDistance();
     void increaseDistance();
 }
