+++ src/main/java/racingcar/service/evaluator/WinnerEvaluator.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface WinnerEvaluator {
     List<RacingCar> determineWinners(List<RacingCar> cars);
-    Long findMaxDistance(List<RacingCar> cars);
-    List<RacingCar> collectWinners(List<RacingCar> cars,Long maxDistance);
+    int findMaxDistance(List<RacingCar> cars);
+    List<RacingCar> collectWinners(List<RacingCar> cars,int maxDistance);
 }
